# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pumba Api",
        "version": "1.0.2"
    },
    "servers": [
        {
            "url": "http://ovz5.j04713753.0n03n.vps.myjino.ru/public/api/",
            "description": "Pumba"
        }
    ],
    "paths": {
...
        "/getChildren": {
            "post": {
                "tags": [
                    "Child"
                ],
                "summary": "Children",
                "description": "Children by api_token",
                "operationId": "children",
                "responses": {
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "User does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "children": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": 0
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": null
                                                    },
                                                    "birth": {
                                                        "type": "string",
                                                        "example": "2022-09-02"
                                                    },
                                                    "api_token": {
                                                        "type": "string"
                                                    },
                                                    "gender": {
                                                        "type": "string",
                                                        "example": "Enum(Boy, Girl, Neutral)"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
        "/getMemories": {
            "post": {
                "tags": [
                    "Memories"
                ],
                "summary": "Memories",
                "description": "Memories by api_token, childId",
                "operationId": "memories",
                "requestBody": {
                    "description": "child id",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "childId"
                                ],
                                "properties": {
                                    "childId": {
                                        "type": "string",
                                        "example": "0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "User does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "memories": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "date": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "note": {
                                                        "type": "string"
                                                    },
                                                    "color": {
                                                        "type": "string"
                                                    },
                                                    "image": {
                                                        "type": "string"
                                                    },
                                                    "childId": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
        "/getProfile": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Profile",
                "description": "Profile by api_token",
                "operationId": "profile",
                "responses": {
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Profile does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "profile": {
                                            "type": "object",
                                            "example": {
                                                "name": "ExampleName",
                                                "email": "example@example.com",
                                                "fcm_token": null,
                                                "api_token": "string",
                                                "avatar": "1669926220.png",
                                                "role": "Enum(Mother, Father, Other)"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/editProfile": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "editProfile",
                "description": "editProfile by api_token, name, avatar, role",
                "operationId": "editProfile",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Example"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "Enum(Mother, Father, Other)"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": "1669926220.png or Base64 or Null(удали поле опционально null/nil)"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "The name has already been taken."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "success": {
                                            "type": "string",
                                            "example": "Profile edited successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/deleteProfile": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "deleteProfile",
                "description": "deleteProfile by api_token",
                "operationId": "deleteProfile",
                "responses": {
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Profile does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "success": {
                                            "type": "string",
                                            "example": "User account deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/setResultQuiz": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "setResultQuiz",
                "description": "setResultQuiz by ages, energy_level, door_type, stuff, time",
                "operationId": "setResultQuiz",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ages",
                                    "energy_level",
                                    "door_type",
                                    "stuff",
                                    "time"
                                ],
                                "properties": {
                                    "ages": {
                                        "type": "string",
                                        "example": "0-1,2-4"
                                    },
                                    "energy_level": {
                                        "type": "string",
                                        "example": "High"
                                    },
                                    "door_type": {
                                        "type": "string",
                                        "example": "Indoor / door"
                                    },
                                    "stuff": {
                                        "type": "string",
                                        "example": "Needed / Not needed"
                                    },
                                    "time": {
                                        "type": "string",
                                        "example": "10-15,15-30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "User does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "success": {
                                            "type": "string",
                                            "example": "Filters added successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}